+++ packages/orq-rc/RELEASES.md
@@ -8769,3 +8769,13 @@ Based on:
 - [typescript v4.9.0-rc.14] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.14] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.14 - packages/orq-rc
+
+## 2026-05-01 06:41:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.10 (2.881.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.15] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.15 - packages/orq-rc